+++ runtime/tests/Test_DLManager.cpp
@@ -24,7 +24,7 @@ using namespace Catalyst::Runtime;
 #ifdef __linux__
 TEST_CASE("Test dummy", "[Third Party]")
 {
-    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>("default");
+    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>();
     std::string file("this-file-does-not-exist.so");
     driver->setDeviceName("DummyDevice");
     REQUIRE_THROWS_WITH(driver->loadDevice(file), Catch::Contains("No such file or directory"));
@@ -32,7 +32,7 @@ TEST_CASE("Test dummy", "[Third Party]")
 
 TEST_CASE("Test error message function not found", "[Third Party]")
 {
-    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>("default");
+    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>();
     std::string file("libm.so.6");
     driver->setDeviceName("DummyDevice");
     REQUIRE_THROWS_WITH(driver->loadDevice(file),
@@ -41,56 +41,50 @@ TEST_CASE("Test error message function not found", "[Third Party]")
 
 TEST_CASE("Test error message if init device fails", "[Third Party]")
 {
-    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>("default");
+    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>();
     std::string file("libm.so.6");
-    REQUIRE_THROWS_WITH(driver->initDevice(file),
-                        Catch::Contains("undefined symbol: defaultFactory"));
+    REQUIRE_THROWS_WITH(driver->initDevice(file, "", ""),
+                        Catch::Contains("undefined symbol: Factory"));
 }
 
 TEST_CASE("Test success of loading dummy device", "[Third Party]")
 {
-    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>("default");
-    std::string file("libdummy_device.so");
-    driver->setDeviceName("DummyDevice");
-    CHECK(driver->initDevice(file));
+    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>();
+    CHECK(driver->initDevice("libdummy_device.so", "DummyDevice", ""));
 }
 #endif
 
-TEST_CASE("Test __rt__device registering a custom device with shots=500 and device=lightning.qubit",
+TEST_CASE("Test __quantum__rt__device_init registering a custom device with shots=500 and "
+          "device=lightning.qubit",
           "[CoreQIS]")
 {
     __quantum__rt__initialize();
 
-    char dev[8] = "rtd_lib";
-    char dev_value[17] = "lightning.qubit";
-    __quantum__rt__device((int8_t *)dev, (int8_t *)dev_value);
+    char dev1[17] = "lightning.qubit";
+    __quantum__rt__device_init((int8_t *)dev1, nullptr, nullptr);
 
-    char dev2[7] = "device";
-    char dev2_value[15] = "backend.other";
-    REQUIRE_THROWS_WITH(__quantum__rt__device((int8_t *)dev2, (int8_t *)dev_value),
-                        Catch::Contains("[Function:__quantum__rt__device] Error in Catalyst "
-                                        "Runtime: Invalid device specification"));
-
-    REQUIRE_THROWS_WITH(__quantum__rt__device((int8_t *)dev, (int8_t *)dev2_value),
+    char dev2[15] = "backend.other";
+    REQUIRE_THROWS_WITH(__quantum__rt__device_init((int8_t *)dev2, nullptr, nullptr),
                         Catch::Contains("cannot open shared object file"));
 
-    REQUIRE_THROWS_WITH(__quantum__rt__device(nullptr, nullptr),
-                        Catch::Contains("Invalid device specification"));
+    REQUIRE_THROWS_WITH(__quantum__rt__device_init(nullptr, nullptr, nullptr),
+                        Catch::Contains("Invalid device library"));
 
     __quantum__rt__finalize();
 
-    REQUIRE_THROWS_WITH(__quantum__rt__device((int8_t *)dev, (int8_t *)dev_value),
+    REQUIRE_THROWS_WITH(__quantum__rt__device_init((int8_t *)dev1, nullptr, nullptr),
                         Catch::Contains("Invalid use of the global driver before initialization"));
 }
 
 #ifdef __device_lightning_kokkos
-TEST_CASE("Test __rt__device registering device=lightning.kokkos", "[CoreQIS]")
+TEST_CASE("Test __quantum__rt__device_init registering device=lightning.kokkos", "[CoreQIS]")
 {
     __quantum__rt__initialize();
 
-    char dev[8] = "rtd_lib";
-    char dev_value[18] = "lightning.kokkos";
-    __quantum__rt__device((int8_t *)dev, (int8_t *)dev_value);
+    char rtd_name[18] = "lightning.kokkos";
+    __quantum__rt__device_init((int8_t *)rtd_name, nullptr, nullptr);
+
+    __quantum__rt__device_release();
 
     __quantum__rt__finalize();
 }