+++ runtime/tests/Test_LightningMeasures.cpp
@@ -176,33 +176,33 @@ TEMPLATE_LIST_TEST_CASE("Expval(NamedObs) test", "[Measures]", SimTypes)
     CHECK(sim->Expval(pz) == Approx(-1.0).margin(1e-5));
 }
 
-TEMPLATE_LIST_TEST_CASE("Expval(NamedObs) shots test", "[Measures]", SimTypes)
-{
-    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
+// TEMPLATE_LIST_TEST_CASE("Expval(NamedObs) shots test", "[Measures]", SimTypes)
+// {
+//     std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
-    // state-vector with #qubits = n
-    constexpr size_t n = 4;
-    std::vector<QubitIdType> Qs;
-    Qs.reserve(n);
-    for (size_t i = 0; i < n; i++) {
-        Qs.push_back(sim->AllocateQubit());
-    }
+//     // state-vector with #qubits = n
+//     constexpr size_t n = 4;
+//     std::vector<QubitIdType> Qs;
+//     Qs.reserve(n);
+//     for (size_t i = 0; i < n; i++) {
+//         Qs.push_back(sim->AllocateQubit());
+//     }
 
-    sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
-    sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
-    sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
-    sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
+//     sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
+//     sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
+//     sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
+//     sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
 
-    ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
-    ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
-    ObsIdType pz = sim->Observable(ObsId::PauliZ, {}, {Qs[1]});
+//     ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
+//     ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
+//     ObsIdType pz = sim->Observable(ObsId::PauliZ, {}, {Qs[1]});
 
-    size_t num_shots = 1000;
+//     size_t num_shots = 1000;
 
-    CHECK(sim->Expval(px, num_shots, {}) == Approx(1.0).margin(5e-2));
-    CHECK(sim->Expval(py, num_shots, {}) == Approx(.0).margin(5e-2));
-    CHECK(sim->Expval(pz, num_shots, {}) == Approx(-1.0).margin(5e-2));
-}
+//     CHECK(sim->Expval(px, num_shots, {}) == Approx(1.0).margin(5e-2));
+//     CHECK(sim->Expval(py, num_shots, {}) == Approx(.0).margin(5e-2));
+//     CHECK(sim->Expval(pz, num_shots, {}) == Approx(-1.0).margin(5e-2));
+// }
 
 TEMPLATE_LIST_TEST_CASE("Expval(HermitianObs) test", "[Measures]", SimTypes)
 {
@@ -258,36 +258,36 @@ TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(NamedObs)) test", "[Measures]", SimTy
     CHECK(sim->Expval(tpz) == Approx(-1.0).margin(1e-5));
 }
 
-TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(NamedObs)) shots test", "[Measures]", SimTypes)
-{
-    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
+// TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(NamedObs)) shots test", "[Measures]", SimTypes)
+// {
+//     std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
-    // state-vector with #qubits = n
-    constexpr size_t n = 4;
-    std::vector<QubitIdType> Qs;
-    Qs.reserve(n);
-    for (size_t i = 0; i < n; i++) {
-        Qs.push_back(sim->AllocateQubit());
-    }
+//     // state-vector with #qubits = n
+//     constexpr size_t n = 4;
+//     std::vector<QubitIdType> Qs;
+//     Qs.reserve(n);
+//     for (size_t i = 0; i < n; i++) {
+//         Qs.push_back(sim->AllocateQubit());
+//     }
 
-    sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
-    sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
-    sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
-    sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
+//     sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
+//     sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
+//     sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
+//     sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
 
-    ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
-    ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
-    ObsIdType pz = sim->Observable(ObsId::PauliZ, {}, {Qs[1]});
-    ObsIdType tpx = sim->TensorObservable({px});
-    ObsIdType tpy = sim->TensorObservable({py});
-    ObsIdType tpz = sim->TensorObservable({pz});
+//     ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
+//     ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
+//     ObsIdType pz = sim->Observable(ObsId::PauliZ, {}, {Qs[1]});
+//     ObsIdType tpx = sim->TensorObservable({px});
+//     ObsIdType tpy = sim->TensorObservable({py});
+//     ObsIdType tpz = sim->TensorObservable({pz});
 
-    size_t num_shots = 1000;
+//     size_t num_shots = 1000;
 
-    CHECK(sim->Expval(tpx, num_shots, {}) == Approx(1.0).margin(5e-2));
-    CHECK(sim->Expval(tpy, num_shots, {}) == Approx(.0).margin(5e-2));
-    CHECK(sim->Expval(tpz, num_shots, {}) == Approx(-1.0).margin(5e-2));
-}
+//     CHECK(sim->Expval(tpx, num_shots, {}) == Approx(1.0).margin(5e-2));
+//     CHECK(sim->Expval(tpy, num_shots, {}) == Approx(.0).margin(5e-2));
+//     CHECK(sim->Expval(tpz, num_shots, {}) == Approx(-1.0).margin(5e-2));
+// }
 
 TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(NamedObs[])) test", "[Measures]", SimTypes)
 {
@@ -319,44 +319,43 @@ TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(NamedObs[])) test", "[Measures]", Sim
     CHECK(sim->Expval(tpxz) == Approx(-1.0).margin(1e-5));
 }
 
-TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(NamedObs[])) shots test", "[Measures]", SimTypes)
-{
-    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
-    std::unique_ptr<TestType> sim0 = std::make_unique<TestType>();
+// TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(NamedObs[])) shots test", "[Measures]", SimTypes)
+// {
+//     std::unique_ptr<TestType> sim = std::make_unique<TestType>();
+//     std::unique_ptr<TestType> sim0 = std::make_unique<TestType>();
 
-    // state-vector with #qubits = n
-    constexpr size_t n = 4;
-    std::vector<QubitIdType> Qs;
-    Qs.reserve(n);
-    for (size_t i = 0; i < n; i++) {
-        Qs.push_back(sim->AllocateQubit());
-        Qs.push_back(sim0->AllocateQubit());
-    }
+//     // state-vector with #qubits = n
+//     constexpr size_t n = 4;
+//     std::vector<QubitIdType> Qs;
+//     Qs.reserve(n);
+//     for (size_t i = 0; i < n; i++) {
+//         Qs.push_back(sim->AllocateQubit());
+//         Qs.push_back(sim0->AllocateQubit());
+//     }
 
-    sim->SetDeviceShots(1000);
-    
-    
-    sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
-    sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
-    sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
-    //sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
+//     sim->SetDeviceShots(1000);
 
-    sim0->NamedOperation("PauliX", {}, {Qs[0]}, false);
-    sim0->NamedOperation("PauliY", {}, {Qs[1]}, false);
-    sim0->NamedOperation("PauliZ", {}, {Qs[3]}, false);
-    //sim0->NamedOperation("Hadamard", {}, {Qs[2]}, false);
-    
+//     sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
+//     sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
+//     sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
+//     //sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
 
-    ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
-    ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
-    ObsIdType tpxy = sim->TensorObservable({px, py});
+//     sim0->NamedOperation("PauliX", {}, {Qs[0]}, false);
+//     sim0->NamedOperation("PauliY", {}, {Qs[1]}, false);
+//     sim0->NamedOperation("PauliZ", {}, {Qs[3]}, false);
+//     //sim0->NamedOperation("Hadamard", {}, {Qs[2]}, false);
 
-    ObsIdType px0 = sim0->Observable(ObsId::PauliX, {}, {Qs[2]});
-    ObsIdType py0 = sim0->Observable(ObsId::PauliY, {}, {Qs[1]});
-    ObsIdType tpxy0 = sim0->TensorObservable({px0, py0});
+//     ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
+//     ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
+//     ObsIdType tpxy = sim->TensorObservable({px, py});
 
-    CHECK(sim->Expval(tpxy, sim->GetDeviceShots(), {}) == Approx(sim0->Expval(tpxy0)).margin(5e-2));
-}
+//     ObsIdType px0 = sim0->Observable(ObsId::PauliX, {}, {Qs[2]});
+//     ObsIdType py0 = sim0->Observable(ObsId::PauliY, {}, {Qs[1]});
+//     ObsIdType tpxy0 = sim0->TensorObservable({px0, py0});
+
+//     CHECK(sim->Expval(tpxy, sim->GetDeviceShots(), {}) ==
+//     Approx(sim0->Expval(tpxy0)).margin(5e-2));
+// }
 
 TEMPLATE_LIST_TEST_CASE("Expval(TensorProd(HermitianObs))", "[Measures]", SimTypes)
 {
@@ -513,31 +512,31 @@ TEMPLATE_LIST_TEST_CASE("Expval(Hamiltonian(NamedObs[])) test", "[Measures]", Si
     CHECK(sim->Expval(hxyz) == Approx(0.2).margin(1e-5));
 }
 
-TEMPLATE_LIST_TEST_CASE("Expval(Hamiltonian(NamedObs[])) shots test", "[Measures]", SimTypes)
-{
-    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
+// TEMPLATE_LIST_TEST_CASE("Expval(Hamiltonian(NamedObs[])) shots test", "[Measures]", SimTypes)
+// {
+//     std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
-    // state-vector with #qubits = n
-    constexpr size_t n = 4;
-    std::vector<QubitIdType> Qs;
-    Qs.reserve(n);
-    for (size_t i = 0; i < n; i++) {
-        Qs.push_back(sim->AllocateQubit());
-    }
+//     // state-vector with #qubits = n
+//     constexpr size_t n = 4;
+//     std::vector<QubitIdType> Qs;
+//     Qs.reserve(n);
+//     for (size_t i = 0; i < n; i++) {
+//         Qs.push_back(sim->AllocateQubit());
+//     }
 
-    sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
-    sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
-    sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
-    sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
+//     sim->NamedOperation("PauliX", {}, {Qs[0]}, false);
+//     sim->NamedOperation("PauliY", {}, {Qs[1]}, false);
+//     sim->NamedOperation("Hadamard", {}, {Qs[2]}, false);
+//     sim->NamedOperation("PauliZ", {}, {Qs[3]}, false);
 
-    ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
-    ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
-    ObsIdType pz = sim->Observable(ObsId::PauliZ, {}, {Qs[1]});
-    ObsIdType hxyz = sim->HamiltonianObservable({0.4, 0.8, 0.2}, {px, py, pz});
+//     ObsIdType px = sim->Observable(ObsId::PauliX, {}, {Qs[2]});
+//     ObsIdType py = sim->Observable(ObsId::PauliY, {}, {Qs[1]});
+//     ObsIdType pz = sim->Observable(ObsId::PauliZ, {}, {Qs[1]});
+//     ObsIdType hxyz = sim->HamiltonianObservable({0.4, 0.8, 0.2}, {px, py, pz});
 
-    size_t num_shots = 1000;
-    CHECK(sim->Expval(hxyz, num_shots, {}) == Approx(0.2).margin(5e-2));
-}
+//     size_t num_shots = 1000;
+//     CHECK(sim->Expval(hxyz, num_shots, {}) == Approx(0.2).margin(5e-2));
+// }
 
 TEMPLATE_LIST_TEST_CASE("Expval(Hamiltonian(TensorObs[])) test", "[Measures]", SimTypes)
 {